+++ examples/tester/esm/eslint.config.ts
@@ -1,0 +1,11 @@
+import { makeConfig } from '@rockpack/codestyle';
+
+const config = makeConfig();
+
+config.push({
+  rules: {
+    'package-json/require-type': 'off',
+  },
+});
+
+export default config;